\name{snowfall-data}
\docType{data}
\alias{config}
\alias{f1}
\alias{f2}
\title{Internal configuration and test data}
\description{
  Internal configuration and test data. Only used for internal setup and
  testing.
}
\usage{
config
f1
f2
}
\format{A matrix containing basic predefined configuration informations.}
\source{}
\references{}
\keyword{datasets}