% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{hi}
\alias{hi}
\title{Start a conversation}
\usage{
hi(who, how = c("shout", "whisper", "asis"))
}
\arguments{
\item{who}{character(1) Name of person to be greeted}

\item{how}{character(1) (optional). How is the person to be greeted. Options
include "shout", "whisper" and "asis".}
}
\value{
A character(1) vector containing appropriate greeting.
}
\description{
This function will produce a friendly greeting, useful in starting 
conversations.
}
\examples{
hi("Craig")
#\dontrun{   # can also use donttest{}
#   hi("Craig", "shout")
# }
}

